created Controller Methods trait to reuse common methods in Purchase Controller and ProductPurchaseController
