+++ app/Http/Controllers/PurchaseController.php
@@ -7,12 +7,13 @@ use App\Models\Purchase;
 use App\Models\Product;
 use App\Models\ProductPurchase;
 use Carbon\Carbon;
-use Illuminate\Support\Facades\DB;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Redirect;
+use App\Traits\ControllerMethods;
 
 class PurchaseController extends Controller
 {
+
+  use ControllerMethods;
+
   // **Muestra la lista de compras**
   public function index()
   {
@@ -116,26 +117,6 @@ class PurchaseController extends Controller
       'sortedProducts' => $this->getSortedProducts(),
       'productsPurchases' => $this->getSortedPurchasesById(),
     ]);
-  }
-
-  // **Obtener las compras de productos ordenadas por id descendente**
-  private function getSortedPurchasesById()
-  {
-    return ProductPurchase::orderBy('id', 'desc')->get();
-  }
-
-  // **Obtener productos ordenados por descripcion**
-  private function getSortedProducts()
-  {
-    return Product::whereNull('deleted_at') //Devuelve los productos que NO han sido eliminados en la tabla 'products'
-      ->orderBy('description')
-      ->get();
-  }
-
-  // **Obtener todos los productos**
-  private function getAllProducts()
-  {
-    return Product::withTrashed()->get(); //Devuelve TODOS los productos esten elimados o no en la tabla 'products'
   }
 
   // **obtener el importe total de una compra**
